ts-avatar-service/requirements.txt
click==7.1.2
dlib==19.21.1
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.2
Werkzeug==1.0.1
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-distro===0.36b0
requests==2.28.1
# opentelemetry-api==1.13.0
# opentelemetry-sdk==1.13.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0
